rename dsl package to sim, extract puppet into own file
