+++ sim/commands.go
@@ -1,4 +1,4 @@
-package dsl
+package sim
 
 import (
 	"context"
