+++ sim/dsl.go
@@ -1,4 +1,4 @@
-package dsl
+package sim
 
 import (
 	"context"
@@ -6,14 +6,12 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
 	"log"
 	"net"
 	"os"
 	"os/exec"
 	"os/signal"
 	"path/filepath"
-	"runtime"
 	"sort"
 	"strconv"
 	"strings"
@@ -24,154 +22,20 @@ import (
 	"golang.org/x/sync/errgroup"
 )
 
+const DefaultShsCaps = "1KHLiKZvAvjbY1ziZEHMXawbCEIM6qwjCDm3VYRan/s="
+type Args struct {
+	Caps        string // global caps setting
+	Hops        int    // global hops setting
+	FixturesDir string // directory containing the spliced ssb-fixtures
+	Testfile    string // path to file containing sim statements
+	Outdir      string // directory where puppet logs & files will be dumped
+	BasePort    int    // starting port used for instantiating the ports used by puppets
+	Verbose     bool	 // produce more output when running (echoes puppet output in realtime, in addition to TAP assertions)
+}
+
 type Process struct {
 	cmd     *exec.Cmd
 	logfile *os.File
-}
-
-type Puppet struct {
-	directory     string
-	feedID        string
-	name          string
-	caps          string
-	secretDir     string
-	omitOffset    bool
-	port          int
-	hops          int
-	seqno         int
-	totalMessages int
-	totalTime     time.Duration
-	slept         time.Duration
-	lastStart     time.Time
-	process       Process // holds cmd & logfile of a running puppet process
-}
-
-func (p Puppet) String() string {
-	return fmt.Sprintf("[%s@%d] %s", p.name, p.seqno, p.feedID)
-}
-
-func (p *Puppet) stopTimer() {
-	if p.lastStart.IsZero() {
-		return
-	}
-	p.totalTime += time.Since(p.lastStart)
-	var zero time.Time
-	p.lastStart = zero
-}
-
-func (p *Puppet) countMessages() error {
-	seqnos, err := queryLatest(p)
-	if err != nil {
-		return err
-	}
-	count := 0
-	for _, seqno := range seqnos {
-		count += seqno.Sequence
-	}
-	p.totalMessages = count
-	return nil
-}
-
-func (p *Puppet) addSleepDuration(d time.Duration) {
-	p.slept += d
-}
-
-func (p *Puppet) start(s Simulator, shim string) error {
-	filename := filepath.Join(s.puppetDir, fmt.Sprintf("%s.txt", p.name))
-	// open the log file and append to it. if it doesn't exist, create it first
-	logfile, err := os.OpenFile(filename, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
-	// io.MultiWriter is golang's equivalent of running unix pipes with tee
-	var writer io.Writer
-	writer = logfile
-	if s.verbose {
-		writer = io.MultiWriter(os.Stdout, logfile)
-	}
-	if err != nil {
-		return TestError{err: err, message: "could not create log file"}
-	}
-	var cmd *exec.Cmd
-
-	// currently the simulator has a requirement that each language implementation folder must contain a sim-shim.sh file
-	// sim-shim.sh contains logic for starting the corresponding sbot correctly.
-	// e.g. reading the passed in ssb directory ($1) and port ($2)
-	shimPath := filepath.Join(s.implementations[shim], "sim-shim.sh")
-	cmd = exec.CommandContext(s.rootCtx, shimPath, p.directory, strconv.Itoa(p.port))
-
-	// the environment variables CAPS and HOPS contains the caps (default: ssb caps) and hops (default: 2) settings for
-	// the puppet, and must be set correctly in each implementation's sim-shim.sh
-	cmd.Env = append(os.Environ(),
-		fmt.Sprintf("CAPS=%s", p.caps),
-		fmt.Sprintf("HOPS=%d", p.hops))
-
-	if s.fixtures != "" && p.usesFixtures() {
-		// pass in LOG_OFFSET and SECRET separately, to allow for using a secret w/ no log.offset.
-		// this allows us to simulate when a peer friend-restores their database using only their secret
-		cmd.Env = append(cmd.Env,
-			fmt.Sprintf("SECRET=%s", filepath.Join(s.fixtures, p.secretDir, "secret")))
-		if !p.omitOffset {
-			cmd.Env = append(cmd.Env,
-				fmt.Sprintf("LOG_OFFSET=%s", filepath.Join(s.fixtures, p.secretDir, "flume", "log.offset")))
-		}
-	}
-
-	cmd.Stderr = writer
-	cmd.Stdout = writer
-	// store cmd & logfile in puppet for use when we shut it down with e.g. the stop command
-	p.process = Process{cmd: cmd, logfile: logfile}
-	err = cmd.Start()
-	if err != nil {
-		return TestError{err: err, message: fmt.Sprintf("failure when creating puppet, see %s for information", filename)}
-	}
-
-	return nil
-}
-
-func (p *Puppet) stop() error {
-	// update the total message count before we stop this puppet
-	err := p.countMessages()
-	if err != nil {
-		taplog(fmt.Sprintf("%s had an error when trying to count db messages (%s)", p.name, err))
-	}
-	cmd, logfile := p.process.cmd, p.process.logfile
-	taplog(fmt.Sprintf("stopping %s (%s)", p.name, p.feedID))
-	// issue an interrupt to the process (allows us to do cleanup in sbots)
-	// Windows doesn't support Interrupt
-	if runtime.GOOS == "windows" {
-		cmd.Process.Signal(os.Kill)
-	} else {
-		cmd.Process.Signal(os.Interrupt)
-	}
-
-	// last resort shutdown
-	go func() {
-		time.Sleep(2 * time.Second)
-		_ = cmd.Process.Signal(os.Kill)
-	}()
-
-	// wait for the process to wrap up
-	err = cmd.Wait()
-	if err != nil {
-		return TestError{err: err, message: fmt.Sprintf("failure when stopping puppet")}
-	}
-	// close the logfile
-	err = logfile.Close()
-	if err != nil {
-		return TestError{err: err, message: fmt.Sprintf("failure when closing logfile")}
-	}
-	p.process = Process{}
-	return nil
-}
-
-func (p Puppet) usesFixtures() bool {
-	return len(p.feedID) > 0 && len(p.secretDir) > 0
-}
-
-func (p Puppet) isExecuting() bool {
-	return p.process != Process{}
-}
-
-func (p *Puppet) bumpSeqno() {
-	p.seqno += 1
 }
 
 // TODO: convert all uses of testError to fmt.Errorf(msg + %w)
@@ -657,30 +521,6 @@ func (s Simulator) getPuppet(name string) *Puppet {
 	return p
 }
 
-func preparePuppetDir(dir string) string {
-	// introduce convention that the output dir is called puppets.
-	// this fixes edgecases of accidentally removing unintended
-	// folders + files
-	if filepath.Base(dir) != "puppets" {
-		dir = filepath.Join(dir, "puppets")
-	}
-	absdir, err := filepath.Abs(dir)
-	if err != nil {
-		log.Fatalln(err)
-	}
-	// remove the puppet dir and its subfolders
-	err = os.RemoveAll(absdir)
-	if err != nil {
-		log.Fatalln(err)
-	}
-	// recreate it
-	err = os.Mkdir(absdir, 0777)
-	if err != nil {
-		log.Fatalln(err)
-	}
-	return absdir
-}
-
 func (s Simulator) monitorInterrupts() {
 	c := make(chan os.Signal)
 	signal.Notify(c, os.Interrupt, syscall.SIGTERM)
@@ -727,7 +567,29 @@ func (s Simulator) exit() {
 	time.Sleep(1 * time.Second)
 }
 
-const DefaultShsCaps = "1KHLiKZvAvjbY1ziZEHMXawbCEIM6qwjCDm3VYRan/s="
+func preparePuppetDir(dir string) string {
+	// introduce convention that the output dir is called puppets.
+	// this fixes edgecases of accidentally removing unintended
+	// folders + files
+	if filepath.Base(dir) != "puppets" {
+		dir = filepath.Join(dir, "puppets")
+	}
+	absdir, err := filepath.Abs(dir)
+	if err != nil {
+		log.Fatalln(err)
+	}
+	// remove the puppet dir and its subfolders
+	err = os.RemoveAll(absdir)
+	if err != nil {
+		log.Fatalln(err)
+	}
+	// recreate it
+	err = os.Mkdir(absdir, 0777)
+	if err != nil {
+		log.Fatalln(err)
+	}
+	return absdir
+}
 
 func Run(args Args, sbots []string) {
 	// validate flag-passed caps key
@@ -761,14 +623,4 @@ func Run(args Args, sbots []string) {
 
 	// once we are done we want all puppets to exit
 	sim.exit()
-}
-
-type Args struct {
-	Caps        string // global caps setting
-	Hops        int    // global hops setting
-	FixturesDir string // directory containing the spliced ssb-fixtures
-	Testfile    string // path to file containing dsl statements
-	Outdir      string // directory where puppet logs & files will be dumped
-	BasePort    int    // starting port used for instantiating the ports used by puppets
-	Verbose     bool
 }
